Converted files for mac

+++ source/game.d
@@ -137,7 +137,7 @@ private double[] getParametersBounds(Parameter param, Game game, Season season) 
 private double getValue(Parameter param, Game game, Season season) {
   Team team = ATTRIBUTES_TEAM[param.name];
   string teamName = getTeamName(game, team);
-  int position = countUntil(season.games, game);
+  long position = countUntil(season.games, game);
   int counter = param.numberOfGames;
   if (param.numberOfGames == 0) {
     counter = 1;
